worker/requirements.txt
noisereduce
librosa
soundfile
faster_whisper
pyannote.audio
pika